Extract SubsetSizer into it's own module so it can be re-used.
