+++ analysis/pfe_methods/BUILD
@@ -1,6 +1,8 @@
 py_library(
     name = "pfe_methods",
     srcs = [
+        "optimal_pfe_method.py",
+        "subset_sizer.py",
         "unicode_range_pfe_method.py",
         "whole_font_pfe_method.py",
     ],
@@ -31,6 +33,32 @@ py_test(
     name = "unicode_range_pfe_method_test",
     srcs = [
         "unicode_range_pfe_method_test.py",
+    ],
+    data = [
+        "//patch_subset:testdata",
+    ],
+    deps = [
+        ":pfe_methods",
+    ],
+)
+
+py_test(
+    name = "optimal_pfe_method_test",
+    srcs = [
+        "optimal_pfe_method_test.py",
+    ],
+    data = [
+        "//patch_subset:testdata",
+    ],
+    deps = [
+        ":pfe_methods",
+    ],
+)
+
+py_test(
+    name = "subset_sizer_test",
+    srcs = [
+        "subset_sizer_test.py",
     ],
     data = [
         "//patch_subset:testdata",
