+++ analysis/pfe_methods/unicode_range_pfe_method.py
@@ -12,28 +12,22 @@ These definitions are based on what Google Fonts uses for their production
 font serving.
 """
 
-import io
 import os
 
 from analysis import request_graph
+from analysis.pfe_methods import subset_sizer
 from analysis.pfe_methods.unicode_range_data import slicing_strategy_loader
-from fontTools import subset
-from woff2_py import woff2
 
 # Cache of which slicing strategy to use per font. Keyed by font name.
 FONT_SLICING_STRATEGY_CACHE = dict()
-
-# Cache of cut and woff2 encoded subset sizes. Keyed by:
-# <font name>:<slicing_strategy>:<subset index>
-SUBSET_SIZE_CACHE = dict()
 
 
 def name():
   return "GoogleFonts_UnicodeRange"
 
 
-def start_session(font_directory, subset_sizer=None):
-  return UnicodeRangePfeSession(font_directory, subset_sizer)
+def start_session(font_directory, a_subset_sizer=None):
+  return UnicodeRangePfeSession(font_directory, a_subset_sizer)
 
 
 def slicing_strategy_for_font(font_id, font_bytes):
@@ -48,41 +42,13 @@ def slicing_strategy_for_font(font_id, font_bytes):
           slicing_strategy_loader.load_slicing_strategy(strategy_name))
 
 
-class SubsetSizer:
-  """Helper class that computes the woff2 encoded size of a font subset."""
-
-  def subset_size(self, cache_key, codepoints, font_bytes):
-    """Returns the size of subset (a set of codepoints) of font_bytes after woff2 encoding."""
-    if cache_key in SUBSET_SIZE_CACHE:
-      return SUBSET_SIZE_CACHE[cache_key]
-
-    subset_bytes = self.subset(font_bytes, codepoints)
-    woff2_bytes = woff2.ttf_to_woff2(subset_bytes)
-
-    final_size = len(woff2_bytes)
-    SUBSET_SIZE_CACHE[cache_key] = final_size
-    return final_size
-
-  def subset(self, font_bytes, codepoints):  # pylint: disable=no-self-use
-    """Computes a subset of font_bytes to the given codepoints."""
-    options = subset.Options()
-    subsetter = subset.Subsetter(options=options)
-    with io.BytesIO(font_bytes) as font_io, \
-         subset.load_font(font_io, options) as font:
-      subsetter.populate(unicodes=codepoints)
-      subsetter.subset(font)
-
-      with io.BytesIO() as output:
-        subset.save_font(font, output, options)
-        return output.getvalue()
-
-
 class UnicodeRangePfeSession:
   """Unicode range PFE session."""
 
-  def __init__(self, font_directory, subset_sizer=None):
+  def __init__(self, font_directory, a_subset_sizer=None):
     self.font_directory = font_directory
-    self.subset_sizer = subset_sizer if subset_sizer else SubsetSizer()
+    self.subset_sizer = a_subset_sizer if a_subset_sizer else subset_sizer.SubsetSizer(
+    )
     self.request_graphs = []
     self.already_loaded_subsets = set()
